Ensure subtitles are always visible after re-enabling or track switching during pause
Currently, since subtitles are read passively/"lazily",
ie. the sub packet queue is only filled as a/v is being demuxed.
So, during pause, the sub packets of a newly enabled sub track
are not available to the decoder until the segment that contains
the start pts of the subtitle is re-demuxed.
I think this was meant to be solved by issuing refresh seeks
(cf. the preroll mechanism) after selecting a track which
seek back 10 secs and grab all packets missing from the packet queue
along the way. However, this doesn’t work in all cases,
because it stops (marks the sub stream as eof) after the first packet
it finds. Similarly, demux_read_packet_async_until is only called once,
even though there might have been multiple packets in the past 10 secs
and so the current subtitles might not be included.

Fix this by forcing the demuxer to always read until the current pts and
by draining the queue completely before moving on to decoding, so the current
subs are sure to be included.