demux: fix eof signaling for lazy streams when force reading
We need to account for the edge case where both the stream is eof and the reader_head
is NULL. In this case we still need to return 0 (wait) if the lazy stream requires it
(i.e. if we’re force reading subs).
This may occur when for example
- the sub demuxer queue contains a packet, so it’s marked as eof by read_packet
- then, dequeue_packet is called twice in quick succession, so that
after the first call, the readerhead is NULL and the stream is still eof
because the demuxer thread hasn’t had time to respond to the
readerhead advancement and read in a new packet.