sub: read and decode unselected subtitle tracks
Usually, packets belonging to currently unselected tracks
are skipped/discarded during demuxing. This means that on track
switches, mpv doesn’t have the new track’s data for the current pts
and needs to seek back and re-demux the last 10 seconds to collect
the new track’s packets. Also, every time a sub track is switched,
the old track is destroyed and the new track is initialized from scratch. This behaviour is very inefficient, especially in these two cases:
- external subs need to be fully re-read from disk into memory and re-decoded every time they’re selected
- when streaming over network, sub track switches take multiple seconds because seeking in network streams is slow

With this commit, ALL subtitle tracks, selected or not, are read, decoded
and cached during demuxing, eliminating the need for refresh seeks
(though they haven’t been removed yet) and the need to reinitialize
the sub decoder for newly selected tracks (meaning the ASS_Track survives
tracks switches and so external files don’t need to be re-read).

However, this obviously wastes some memory and cpu in the background, so I should make it configurable at some point.