WIP: sub: read and decode unselected tracks
Usually, packets belonging to currently unselected tracks
are skipped/discarded during demuxing. This means that on track
switches, mpv doesn’t have the new track’s data for the current pts
and needs to seek back and re-demux the last 10 seconds to collect
the new track’s packets. Also, every time a sub track is switched,
the old track is destroyed and the new track is initialized from scratch.
This behaviour causes isses in some cases:
- the case that f8e50dc fixed with an ugly hack, which can be made
obsolete with this commit (but hasn’t been yet)
- external subs need to be re-read from disk into memory and re-decoded
every time they’re selected, which is pretty inefficient
- when streaming over network, sub track switches take multiple seconds
because seeking in network streams is slow

With this commit, ALL subtitle tracks, selected or not, are read, decoded
and cached during demuxing, eliminating the need for refresh seeks
(though they haven’t been removed yet) and the need to reinitialize
the sub decoder for newly selected tracks (meaning the ASS_Track survives
tracks switches and so external files don’t need to be re-read).

This unfortunately breaks secondary subs because all sub tracks
are initialzed as primary subs and there is no mechanism yet
to change a track’s order on the fly except by reinitializing it
(which I want to avoid).

This shit is obviously pretty wasteful outside of the scenarios
listed above, so I should make it configurable at some point.