scripting: add second C plugin entrypoint for Windows support
The previous unix-only C plugin system worked by leaving references to
mpv symbols in the plugin unresolved at compile time and resolving them
with the executable’s symbols at runtime. This did not work
on Windows due to the explicit nature of the table of imported symbols.
This commit adds a second entrypoint that takes a struct of all libmpv API
functions as a parameter to work around the restrictions in PE and enable
Windows support.

Since the new entrypoint should be supported by all platforms,
the compiler checks in the build system are dropped.
If there is a non-win32 platform that doesn’t support -rdynamic,
this might lead to issues, but who cares.

scripting: add missing apis to mpapi

Pretty fucking ugly but who cares it seems to work