player, demux: add command that renders a thumbnail from the demuxer cache
Essentially just overlay-add except the user
doesn’t supply the image but the pts and if it’s
in the demuxer cache, mpv decodes the enclosing
keyframe sequence and adds the decoded frame as an overlay.

The purpose of this is to enable some sort of
rudimentary thumbnailing on network streams (since
thumbfast.lua relies on spawning a secondary mpv instance,
loading the file and seeking in it, which would be
very slow/waste a lot of bandwidth for streams, unless you enabled
the cache in the slave mpv instance, but that would waste
a lot memory and consume twice the bandwidth still).

For local files this is much slower than thumbfast
because it’s implemented very naively and lacks
all the optimizations of mpv’s regular seeking
(which means it wastes a lot cpu cycles.. pick
your poison I guess).